[UNSTABLE] weather tick update
I think that we have most of the important details down, but I need to find a different home for the tick updates. I accidentally assumed that "handle" had access to full_feed, but it can actually only look at the feed given to it. It will be clumsy, but I think the best solution (especially since this is still a demonstration objective) is to put that code in the difference-calculator function. Then we can try re-running.