Added save-to-csv feature
Now we're cooking with peanut oil. The next step is to create an auto-save feature which will write CSVs to a log directory (month folders within year folders; each CSV is named after the full date for redundancy reasons) after every AUTOSAVE_INTERVAL seconds, which will remain a constant at the top of shared, so that future users can easily modify it.

The next step after that is to create a more intelligent launch process. It looks for log.txt and uses today's date to get a CSV that it should resume from.