Auto-save added, with nested directories
We've created an auto-save feature which writes the two WS DFs to CSVs in a record directory (month folders within year folders) after every AUTOSAVE_INTERVAL seconds. AUTOSAVE_INTERVAL is a constant at the top of shared, so that future users can easily modify it.

The next step is to create a more intelligent launch process. It looks for log.txt and uses today's date to get a CSV that it should resume from. log.txt will tell the launcher whether the program quit gracefully or whether the system crashed.

Unfortunately, we're implementing this a little out of order, because currently it's not even possible to quit gracefully. I always have to force quit the tool. Maybe we should address that first?